+++ ktse/src/main/kotlin/net/kigawa/keruta/ktse/persist/db/table/QueueTable.kt
@@ -4,9 +4,9 @@ import org.jetbrains.exposed.sql.Table
 import org.jetbrains.exposed.sql.kotlin.datetime.datetime
 
 object QueueTable: Table("queue") {
-    val id = integer("id").autoIncrement()
+    val id = long("id").autoIncrement()
     val name = varchar("name", 50)
-    val providerId = integer("provider_id").references(ProviderTable.id)
+    val providerId = long("provider_id").references(ProviderTable.id)
     val setting = text("setting")
     val createdAt = datetime("created_at")
     override val primaryKey = PrimaryKey(id)
