+++ ktse/src/main/kotlin/net/kigawa/keruta/ktse/persist/db/table/QueueUserTable.kt
@@ -4,8 +4,8 @@ import org.jetbrains.exposed.sql.Table
 import org.jetbrains.exposed.sql.kotlin.datetime.datetime
 
 object QueueUserTable: Table("queue_user") {
-    val queueId = integer("queue_id").references(QueueTable.id)
-    val userId = integer("user_id").references(UserTable.id)
+    val queueId = long("queue_id").references(QueueTable.id)
+    val userId = long("user_id").references(UserTable.id)
     val createdAt = datetime("created_at")
     override val primaryKey = PrimaryKey(queueId, userId)
 }
